Fix playlists getting cleared when they are edited
As of commit 7586dd4b30cfe5ca29e5bfacd562ef14ed285c22,
when playlist entries were loaded  their child songs were
assigned incorrect ID's. As a result, because of the playlist
editing implementation, when moving items or removing
individual items, the entire playlist would be cleared in the
MediaStore.

This has been fixed by assigning the correct ID to the
Playlist entries when they are loaded.