Change skip previous gesture behavior
When swiping to the right to skip previous, the queue
position will now always decrement to the previous
song, regardless of the current seek position. If the
gesture is preformed at the first song in the queue,
then it will wrap to the last song in the queue if
repeat all is enabled, otherwise the current song will
be restarted from the beginning.