Make rule filtering faster
With the current implementation of LocalMusicStore, calls to getSongs(Album|Artist) are implemented by reading the MediaStore with a Cursor. This is inefficient because opening Cursors is expensive and these values are already stored as properties on the song model object. This commit changes the implementation of the Album and Artist rule so that instead of getting the song contents from the MediaStore with LocalMusicStore, they manually perform a linear search through the song library comparing whether or not each song should be included or not based on its album or artist id.