Fix unexpected behavior when reordering queue items
Fixes an issue where reordering the items in the queue would seemingly not have any affect. This was caused by the ViewModels associated with each itemView not updating their index when an item in a list was reorganized. To solve this issue, DragDropAdapter now calls notifyDataSetChange after each drop to make sure that ViewModels are updated  with correct data