Fix album grid frame colors not updating
Fixes an issue where palette generation in AlbumViewModel would not update the view in release builds. This was because AlbumViewModel uses an ObjectAnimator to change the value of an ObservableInt. When Proguard obfuscated property names, the set(int) method could no longer be accessed by ObjectAnimator because it uses reflection to access JavaBean-style setters and could not find the renamed method. This commit resolves the issue by preserving the names of all members in any class that extends BaseObservable (including ObservableInt) when Proguard runs.