Move Multi-Repeat setting into a process-specific SharedPreferences
Because the Multi-Repeat value needs to always be writeable from the remote process, this PR creates a separate RemotePreferenceStore to hold the value. This is backed by a different SharedPreferences file so that there are no conflicts when using preferences among two processes. This commit also implements the Multi-Repeat behavior on the service and allows its current value to be get and set with PlayerController.