Fix IllegalStateException while dragging list items
Fixes an issue where dragging an item could cause an IllegalStateException if the user lets go of the item after dragging it offscreen. This was caused by notifying the RecyclerView that the data set had changed while it was being laid out or scrolling, and is resolved by delaying the call to `notifyDataSetChanged()`.