Remove "Add Themed Icon to Launcher" setting
In the implementation of App Shortcuts by Google, only the default
launcher icon may have app shortcuts. Any custom launcher icons will
continue to work, but will not have the ability to open the shortcut menu.
Since this is an unintuitive detail to explain to users, and the setting
is not frequently used, it is easier to remove the ability all together.

Users seeking this feature may still continue to use Jockey's alternate
icons through its icon pack if they are using a custom launcher.