Fix control buttons clipping in MultiWindow
When opening Jockey in multi-window on a phone (e.g. the Google Pixel XL), Jockey's skip buttons will be clipped when the phone is in portrait when Jockey is allocated to 1/3 or 1/2 of the screen. To fix this, there are now two spacer views that have their own weight to push the buttons closer together (rather than having a preset constant width). This weight has to be tweaked depending on the screen size, as the same weight won't work for a tablet as it does for a phone.