Fix unusable player control width in tablet multiwindow
When running Jockey on a large tablet in multiwindow (e.g. the Pixel C), allocating Jockey to take 2/3 of the screen in landscape orientation would cause the player control and now playing queue width to be so small that it is virtually unusable. To fix this, there is now a preset width for the player controls, and the album artwork uses the rest of the available screen area (rather than this being determined vice-versa)