Fix incorrect padding at bottom of queue list
Due to changes in the RecyclerView library, the previous implementation of the queue padding (having a single view whose height changes) no longer works. To resolve this, multiple spacers are shown instead. The height of each spacer is the same as the height of a list item (except for the first one, which is the remainder). To update the padding, we show and hide the number of shown spacers to set the total amount of spacing according to whatever song is playing.