Fix queue position updates not scrolling queue
Every time the queue index is updated, the queue fragment previously would trigger a dataSetChanged event. Because of changes to RecyclerView, this would invalidate all of the views currently displayed. This invalidation reset the indices of views in the adapter to NO_POSITION, which is a value needed by shouldScrollToNowPlaying(). To fix this, the queue is only changed when it actually updates.