Fix incorrect URI path parsing
The previous implementation of NowPlayingActivity.onNewIntent converted a URI to a file path by calling Uri.getPath(). This works only if the given URI is a file URI, which is not guaranteed to be the case. This commit resolves this issue by looking up the file path of the song in the MediaStore from the URI passed in through the intent.