Fix intents not being processed correctly on cold starts
When processing "Play with Jockey" intents, if the player service was not already running, the update would be ignored. This commit resolves this issue by correctly waiting for the service to bind (if it is not already running) before applying the change. The current implementation adds a callback for when the service is bound so that the queue setup can be done at the correct time.