Fix potential startup crash
Fixes a potential crash where the service binder may be returned to the main thread before the service has finished its call to onCreate(). Previously, this would cause accesses and commands to the service to throw NPEs to the main thread. This implementation simply drops all incomming commands if the service has not finished its onCreate call before data is accessed.