Improve player state restoration
Modifies the player state saving and restoration implementation by moving its core logic to the MusicPlayer class and performing the transactions over IPC.

Because this implementation is in the remove service and contains the seek position, it is sent as a single to prevent constant expensive transactions to update the current position field. Since this is designed to be used as an undo implementation, this should not pose any limitations for classes that rely on this field.