Send minor updates from player service to application
In the previous implementation of widgets, it was possible for the
widget's state to become invalid after altering the playback state from
the main application. This commit resolves this issue by sending update
broadcasts for every single state change. To avoid unnecessary work,
updates that are minor (i.e. were caused by the user) have an extra flag
set to true in the intent so that they can be filtered and ignored.