Fix duration initializing to 0 on cold start
Fixes an issue where seek position appears invalid after a cold start. This is because ExoPlayer does not resolve a duration until a song has finished loading. The current workaround is to check if ExoPlayer returns an invalid duration. If it does return an invalid duration, it returns the last valid duration. If there isn't one, it defaults to the duration embedded in the song object. This may be inaccurate since it's pulled from the MediaStore, but should work most of the time.