Improve startup performance
Duplicate calls to bindService are now ignored until at least 500ms have passed since the last request to start the service. This improves the startup performance because there is a delay before the service isstarted. When fields are subscribed to, ServicePlayerController ensuresthat the service is started. Every time this happens, it will register an additional callback to listen for service connected events. This means that every field would be invalidated once for every field that has been subscribed to.