Fix snackbars not appearing reliably
Because only one Snackbar can be visible at a time per Activity, Snackbars could previously not be visibile (since they could appear in the hidden view) on pages with a collapsing now playing page. This commit fixes this issue by ensuring that snackbars are only shown in if their container is visible.