Prevent music from playing when Jockey starts
Because headphone connection broadcasts are sticky intents, registering the headphone connection broadcast receiver would cause Jockey to always start playing music when it is launched if headphones are connected (and the resume playback on headphone connect preference is enabled). This commit resolves this issue by ignoring the initial sticky broadcast.