Fix colliding broadcast permission names
Previously it was impossible for two versions of Jockey to be installed simultaneously (e.g. the debug and release verison), because no two applications are allowed to declare the same permission. This commit makes the permission name generated at compile time so that it's bound to the package name, which prevents this collision.

The permission name also changed slightly so that if someone does have a second version of Jockey installed (like a custom flavor or a debug version), they won't have to uninstall/update/recompile the second version.