Fix song durations not updating when a new song starts
Fixes a UI issue where changing a song to an unloaded song would not update the duration shown on the UI because it isn't fetched in time by ExoPlayer to be read with the initial UI update. A second UI update is dispatched if the first one was sent without a correct duration once the duration is loaded by ExoPlayer.