Refactor MediaBrowser implementation
Refactors the previous implementation of MediaBrowserHelper into a set of classes that represent the hierarchy level-by-level instead of in its entirety. This simplifies the necessary steps to integrate new content types in the MediaBrowser and makes the overall implementation less prone to regressions.

Additionally, playlists contents can now be browsed