Fix mismanagement of service bindings
Bindings are now linked to the lifecycle of application components with a new PlayerController.Binding abstraction. ServicePlayerController will now release the service binding when there are no active Activities and rebind to it when an Activity enters the foreground. This is in line with the bound services documentation and examples from Google, and prevents bindings from randomly restarting the player service when the app isn't running.