Fix Android Auto integration not working if Jockey was closed
When using Android Auto after Jockey had previously been stopped, Android Auto would fail to connect to Jockey and would display "Jockey doesn't appear to be working right now." This was caused by an issue where the main application process was kept in a cached state that retained the old media session token. Because of this error, the browser service would immediately offer the dead media session to Android Auto instead of waiting for the service to rebind and request the new media session token.

The fix here is to simply forget the previous media session token when the service is unbound. If needed it will be fetched from the service again, which will either return the previous value, or the new value if the service was recreated.