Use target fragments instead of tag lookups
This fixes an issue where the sleep timer and multi-repeat value pickers would not work because they were unable to send values back to the calling fragment. This was failing because they were trying to look up a fragment with a specific tag, and the requesting fragment did not have a tag specified.

Instead of adding a tag to the fragments, this commit fixes the issue by replacing the tag lookup with the built-in target fragment mechanism, which removes the requirement that fragments need to have a tag for the result to get delivered to them.