Fix potential UnsupportedOperationException
Fixes an issue where setting the queue as an unmodifiable list (like the list of all songs) might cause an UnsupportedOperationException to be thrown if another component got a reference to the list and attempted to modify it.

An example of when this happens is if you play all songs in the library and reorder an item in the queue from the now playing page.